Implement category fetching and state updates
This commit has introduced the 'name' field to the initialState object in 'categorieSlice.ts', and updated the fetching categories function to update the 'name' state. Additionally, 'CategorieBar.tsx' was modified to make use of the Redux state for categories instead of static data. The sidebar now supports navigation to the home route by clicking on the site logo in 'HamburgerSideBar.tsx'.