Add search functionality to the application
This commit includes the creation of a new Redux slice for search terms using the Redux toolkit. Changes were made to the Searchbar component to dispatch an action that sets the search term in the Redux state. Furthermore, the application's state was updated to include the newly created searchReducer. The search term state is read in the Home component for future implementation of search result handling.