Implement checkout functionality and improve order creation process
This commit introduces checkout functionality in the ShoppingCart component which enables user to place an order. It introduces 'createOrder' action using redux toolkit's 'createAsyncThunk', which posts the user's cart items to the backend. Code modifications to the /shop/order route are also present to handle the data. Moreover, this update checks the validity of user IDs and article IDs, converting them to ObjectId for MongoDB.