Refactor data structures and enhance user orders tracking
This commit includes the refactoring of data structures in orderSlice and productSlice. Additionally, it includes modifications in Sidebar and Orders components to implement a more efficient tracking system for user orders, making use of the fetchOrders dispatch and the React useState and useEffect hooks. Changes also enhance readability by replacing 'orderNumber' and 'user_id' with 'orderNr' and 'userId' respectively.