Enhance OrderDetails component with filteredProducts and total calculation
The OrderDetails component has been updated to fetch product data and filter products that are associated with the order. Hence, a detailed list of products is displayed in each order using hook useEffect. Also, a total sum of the money spent on the order is calculated with a reduce method, offering users added convenience in tracking their purchasing.