Add ArticleComment component and refactor Redux state structure
Implemented a new ArticleComment component for displaying individual comments. Additionally, the Redux state structure for comments and reviews was refactored from an object mapping to a simple array structure. This provides a more intuitive and efficient way of handling and retrieving comments on the front end. Error handling was expanded in reviewSlice and ArticleInfo components as well.