Implement fetch users feature in ArticleComments component
This update introduces the fetching of user data in the ArticleComments component. Also, a new 'fetchUsers' async action was added in the 'authSlice.ts' file. A related API endpoint '/shop/users/' was created in the 'routes.js' file in server routes. Now, the user's last name is also being displayed in the ArticleComments component, along with their comment and user ID.